pyproject.toml
[tool.poetry]
name = "data-base-hh-api"
version = "0.1.0"
description = ""
authors = ["NikitinSemen <pr.pypblrka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
